Create F5 facts
Hi, 

            This script is supposed to parse F5 load balancers and copy certain facts such hostname and OS type of the devices into newly created csv files. The script runs without problems. It's just not able to pick out the appropriate details from the load balancers and put them into a csv file. In the host inventory file, I have just added the IPs of the F5 load balancers under [f5] and the username and password under [f5:vars]. The error I get each time I execute the script is:

fatal: [10.244.10.23]: FAILED! => {"msg": "The task includes an option with an undefined variable. The error was: 'ansible_iostype' is undefined\n\nThe error appears to be in '/etc/ansible/f5.yml': line 30, column 17, but may\nbe elsewhere in the file depending on the exact syntax problem.\n\nThe offending line appears to be:\n\n    - debug: var=facts\n                ^ here\n\nThere appears to be both 'k=v' shorthand syntax and YAML in this task. Only one syntax may be used.\n"} 

This error repeats for the device IP mentioned as server as well as all the devices mentioned in the host file under F5. Even though ansible_iostype is the undefined variable, the variables mentioned before as well as after that also don't show on the csv  file.